code/requirements.txt
Flask-JWT-Extended
Flask-RESTFul
Flask-SQLAlchemy